+++ entry.html
@@ -12,6 +12,9 @@
     <button id="saveBtn">Save Note</button>
     <ul id="notesList"></ul>
   </div>
-  <script src="entry.js"></script>
+  <script type="module" src="features/entry.js"></script>
+  <script type = "module" src="features/saveNotes.js"></script>
+  <script type = "module" src="features/displayNotes.js"></script>
+  <script type = "module" src="features/deleteNotes.js"></script>
 </body>
 </html>
