+++ manifest.json
@@ -10,6 +10,9 @@
     "background": {
       "service_worker": "background.js"
     },
-    "permissions": ["storage"]
+    "permissions": ["storage"],
+    "content_security_policy": {
+    "extension_pages": "script-src 'self' 'wasm-unsafe-eval'; object-src 'self'"
+    }
   }
